ARTEMIS-3850 Eliminate Page Soft Cache. Add Option to read messages based on sizing.
(cherry picked from commit 342565d4c1475e974c5cf874ad419842f037e976)

+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/management/impl/AddressControlImpl.java
@@ -382,7 +382,7 @@ public class AddressControlImpl extends AbstractControl implements AddressContro
    }
 
    @Override
-   public int getNumberOfPages() {
+   public long getNumberOfPages() {
       if (AuditLogger.isBaseLoggingEnabled()) {
          AuditLogger.getNumberOfPages(this.addressInfo);
       }
